Reorder Item.elm
The first 119 lines are now code that would exist in the user's program even if they didn't use this forms library.

Lines 135-435 (300 lines total) are form-related code.

So you can have a complete form for a complex type that takes 100ish lines to define in just 300 lines of additional code.